chore(api): rename APIEndpointInterceptorsConfig
